NO-JIRA - fixing up some broken tests
Some tests were added to a class that was extended, Ive moved these into their own class
